Changed Id Datatype from Long to Integer

+++ src/main/java/com/example/FilmBuffs/dao/FilmRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FilmRepository extends CrudRepository<Film, Long> {
+public interface FilmRepository extends CrudRepository<Film, Integer> {
 
     Film findByTitle(String title);
